+++ inc/console/application.class.php
@@ -355,7 +355,7 @@ class Application extends BaseApplication {
 
       $_SESSION['glpilanguage'] = $lang;
 
-      Session::loadLanguage();
+      Session::loadLanguage('', $this->usePlugins());
    }
 
    /**
@@ -383,6 +383,8 @@ class Application extends BaseApplication {
       }
 
       $plugin = new Plugin();
+      $plugin->init();
+
       $plugins_list = $plugin->getPlugins();
       if (count($plugins_list) > 0) {
          foreach ($plugins_list as $name) {