+++ install/update_941_942.php
@@ -31,37 +31,41 @@
  */
 
 /**
- * Update from 9.4.0 to 9.4.1
+ * Update from 9.4.1 to 9.4.2
  *
  * @return bool for success (will die for most error)
 **/
-function update940to941() {
+function update941to942() {
    global $DB, $migration;
 
    $updateresult     = true;
 
    //TRANS: %s is the number of new version
-   $migration->displayTitle(sprintf(__('Update to %s'), '9.4.1'));
-   $migration->setVersion('9.4.1');
+   $migration->displayTitle(sprintf(__('Update to %s'), '9.4.2'));
+   $migration->setVersion('9.4.2');
 
-   /** Add a search option for profile id */
-   $migration->addPostQuery($DB->buildUpdate(
-      'glpi_displaypreferences',
-      [
-         'num' => '5'
-      ],
-      [
-         'num' => '2',
-         'itemtype' => 'Profile'
-      ]
-   ));
+   /* Remove trailing slash from 'url_base' config */
+   $migration->addPostQuery(
+      $DB->buildUpdate(
+         'glpi_configs',
+         [
+            'value' => new \QueryExpression(
+               'TRIM(TRAILING ' . $DB->quoteValue('/') . ' FROM ' . $DB->quoteName('value') . ')'
+            )
+         ],
+         [
+            'context' => 'core',
+            'name'    => 'url_base'
+         ]
+      )
+   );
+   /* /Remove trailing slash from 'url_base' config */
 
    /** Fix URL of images inside ITIL objects contents */
-   // There is an exact copy of this process in "update941to942()".
-   // First version of this migration was working
+   // This is an exact copy of the same process used in "update940to941()" which was working
    // on MariaDB but not on MySQL due to usage of "\d" in a REGEXP expression.
-   // It has been fixed here for people who had not yet updated to 9.4.1 but have been put there
-   // for people already having updated to 9.4.1.
+   // It has been fixed there for people who had not yet updated to 9.4.1 but have to
+   // be put back here for people already having updated to 9.4.1.
    $migration->displayMessage(sprintf(__('Fix URL of images in ITIL tasks, followups ans solutions.')));
 
    // Search for contents that does not contains the itil object parameter after the docid parameter
@@ -136,12 +140,6 @@ function update940to941() {
       }
    }
    /** /Fix URL of images inside ITIL objects contents */
-
-   // Create a dedicated token for rememberme process
-   if (!$DB->fieldExists('glpi_users', 'cookie_token')) {
-      $migration->addField('glpi_users', 'cookie_token', 'string', ['after' => 'api_token_date']);
-      $migration->addField('glpi_users', 'cookie_token_date', 'datetime', ['after' => 'cookie_token']);
-   }
 
    // ************ Keep it at the end **************
    $migration->executeMigration();